2.4.0: Staff member week status API - fixes

+++ src/main/java/com/bdev/hogwarts_api/service/staff_member/week_status/StaffMemberWeekStatusStorageService.kt
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Service
 interface StaffMemberWeekStatusStorageService {
     fun getAllForStaffMember(staffMemberLogin: String): List<StaffMemberWeekStatus>
     fun set(staffMemberWeekStatus: StaffMemberWeekStatus)
-    fun delete(staffMemberWeekStatusId: StaffMemberWeekStatusId)
+    fun delete(staffMemberWeekStatusId: StaffMemberWeekStatusId): Boolean
 }
 
 @Service
@@ -31,11 +31,21 @@ class StaffMemberWeekStatusStorageServiceImpl @Autowired constructor(
         )
     }
 
-    override fun delete(staffMemberWeekStatusId: StaffMemberWeekStatusId) {
-        staffMemberWeekStatusDao.delete(
-            StaffMemberWeekStatusDtoConverter.convertId(
-                dto = staffMemberWeekStatusId
-            )
-        )
+    override fun delete(staffMemberWeekStatusId: StaffMemberWeekStatusId): Boolean {
+        return StaffMemberWeekStatusDtoConverter.convertId(
+            dto = staffMemberWeekStatusId
+        ).let { idModel ->
+            if (staffMemberWeekStatusDao.exists(idModel)) {
+                staffMemberWeekStatusDao.delete(
+                    StaffMemberWeekStatusDtoConverter.convertId(
+                        dto = staffMemberWeekStatusId
+                    )
+                )
+
+                true
+            } else {
+                false
+            }
+        }
     }
 }
